relocate some code
